package/pkg-generic: ignore error when chmod'ing extracted source
The chmod command included in GNU coreutils 9.0 will exit with 1 when in
a recursive a symlink is met and the mode of both itself and its target
is retained; however this situation is usually expected (such kind of
symlink exists in, e.g., a newly extracted tarball of Linux kernel),
thus not a failure.

Considering this chmod call is only kind of vanishing, ignore error for
it.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>